refactor: alter name app
