+++ eject/urls.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 from django.urls import path
-from eject_app.views import ahome
+from home.views import ahome
 
 from django.conf import settings
 from django.conf.urls.static import static
