+++ home/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class EjectAppConfig(AppConfig):
+class HomeAppConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "eject_app"
+    name = "home"
