+++ home/templates/eject_app/home.html
